+++ Amethyst/Errors/CannotNegateArgument.cs
@@ -1,0 +1,10 @@
+using Geode.Errors;
+using System;
+
+namespace Amethyst.Errors
+{
+    public class CannotNegateArgument(string arg) : GeodeError($"Cannot negate argument \"{arg}\" in target selector")
+    {
+        
+    }
+}