+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 * Allow for duplicate arguments in target selectors.
+* Added target selector argument negations.
 * Removed `.mcfunction` dumping on debug builds.
 * Added `--dump-cmds` to the CLI.
+* Added `--version` to the CLI.